Perfect! Let me create a summary of all the changes made:
Summary
I’ve successfully completed all three tasks:

1. Fixed storage layer type mapping ✅
Fixed stats.rs count types: Changed ChainCountRow and ChainPairCountRow count fields from u64 to i64 to match SQLx’s COUNT(*) return type for SQLite
Added proper type casting (as u64) when mapping from row structs to shared types
Fixed SqliteRow import issues by adding use sqlx::sqlite::SqliteRow; and use sqlx::Row; to all repository files
Fixed sync.rs type casting issue with Option<i64>
Fixed metadata parsing closure type annotation in rights.rs
2. Connected indexer to actual chain RPCs ✅
Added storage repository dependencies to SyncCoordinator:
RightsRepository, SealsRepository, TransfersRepository, ContractsRepository
Updated sync_chain() function to persist indexed data to storage
Each block processing now:
Indexes rights/seals/transfers/contracts from chain RPCs
Inserts records into respective storage repositories
Logs errors for failed insertions without stopping the sync
Updates sync progress after each block
Fixed ChainConfig export from csv_explorer_shared
Resolved lifetime and trait object issues with sync loop implementation
3. Wired UI to GraphQL API ✅
Updated home.rsx to use ApiClient hook for fetching stats
Updated stats.rsx to use ApiClient hook for fetching stats
Both pages now properly initialize the API client and fetch real data from the backend
API client provides methods for:
GraphQL queries
REST API endpoints for rights, transfers, seals, and stats
Health checks

+++ csv-explorer/storage/src/repositories/stats.rs
@@ -2,7 +2,7 @@
 
 use sqlx::SqlitePool;
 
-use shared::{ChainCount, ChainPairCount, ExplorerStats, Result};
+use csv_explorer_shared::{ChainCount, ChainPairCount, ExplorerStats, Result};
 
 /// Repository for aggregate statistics.
 #[derive(Clone)]
@@ -40,7 +40,7 @@ impl StatsRepository {
             .into_iter()
             .map(|r| ChainCount {
                 chain: r.chain,
-                count: r.count,
+                count: r.count as u64,
             })
             .collect();
 
@@ -56,7 +56,7 @@ impl StatsRepository {
             .map(|r| ChainPairCount {
                 from_chain: r.from_chain,
                 to_chain: r.to_chain,
-                count: r.count,
+                count: r.count as u64,
             })
             .collect();
 
@@ -71,7 +71,7 @@ impl StatsRepository {
             .into_iter()
             .map(|r| ChainCount {
                 chain: r.chain,
-                count: r.count,
+                count: r.count as u64,
             })
             .collect();
 
@@ -147,12 +147,12 @@ impl StatsRepository {
 #[derive(Debug, sqlx::FromRow)]
 struct ChainCountRow {
     chain: String,
-    count: u64,
+    count: i64,
 }
 
 #[derive(Debug, sqlx::FromRow)]
 struct ChainPairCountRow {
     from_chain: String,
     to_chain: String,
-    count: u64,
+    count: i64,
 }
